requirements.txt
google-genai
azure-identity
azure-storage-file-datalake
azure-keyvault-secrets
pandas
python-dotenv
requests
pyarrow
confluent-kafka>=2.3.0